requirements.txt
ultralytics==8.3.0
streamlit>=1.25
torch>=2.0
Pillow
opencv-python-headless
